+++ packages/core/src/physics/index.ts
@@ -1,11 +1,11 @@
-export { HitResult } from "./HitResult";
-export { PhysicsManager } from "./PhysicsManager";
-export { PhysicsMaterial } from "./PhysicsMaterial";
 export { CharacterController } from "./CharacterController";
-export * from "./shape";
-export * from "./joint";
-export * from "./enums";
-
 export { Collider } from "./Collider";
+export { CollisionDetectionMode, DynamicCollider, DynamicColliderConstraints } from "./DynamicCollider";
+export { HitResult } from "./HitResult";
+export { PhysicsMaterial } from "./PhysicsMaterial";
+export { PhysicsScene } from "./PhysicsScene";
 export { StaticCollider } from "./StaticCollider";
-export { DynamicCollider, CollisionDetectionMode, DynamicColliderConstraints } from "./DynamicCollider";
+export * from "./enums";
+export * from "./joint";
+export * from "./shape";
+
