+++ packages/core/src/physics/joint/FixedJoint.ts
@@ -1,5 +1,5 @@
 import { Collider } from "../Collider";
-import { PhysicsManager } from "../PhysicsManager";
+import { PhysicsScene } from "../PhysicsScene";
 import { Joint } from "./Joint";
 
 /*
@@ -12,6 +12,6 @@ export class FixedJoint extends Joint {
   override _onAwake() {
     const collider = this._collider;
     collider.collider = this.entity.getComponent(Collider);
-    this._nativeJoint = PhysicsManager._nativePhysics.createFixedJoint(collider.collider._nativeCollider);
+    this._nativeJoint = PhysicsScene._nativePhysics.createFixedJoint(collider.collider._nativeCollider);
   }
 }
