+++ packages/core/src/physics/shape/BoxColliderShape.ts
@@ -1,7 +1,7 @@
 import { ColliderShape } from "./ColliderShape";
 import { IBoxColliderShape } from "@galacean/engine-design";
 import { Vector3 } from "@galacean/engine-math";
-import { PhysicsManager } from "../PhysicsManager";
+import { PhysicsScene } from "../PhysicsScene";
 
 /**
  * Physical collider shape for box.
@@ -24,7 +24,7 @@ export class BoxColliderShape extends ColliderShape {
 
   constructor() {
     super();
-    this._nativeShape = PhysicsManager._nativePhysics.createBoxColliderShape(
+    this._nativeShape = PhysicsScene._nativePhysics.createBoxColliderShape(
       this._id,
       this._size,
       this._material._nativeMaterial
