+++ packages/core/src/physics/shape/SphereColliderShape.ts
@@ -1,6 +1,6 @@
 import { ColliderShape } from "./ColliderShape";
 import { ISphereColliderShape } from "@galacean/engine-design";
-import { PhysicsManager } from "../PhysicsManager";
+import { PhysicsScene } from "../PhysicsScene";
 
 /**
  * Physical collider shape for sphere.
@@ -24,7 +24,7 @@ export class SphereColliderShape extends ColliderShape {
 
   constructor() {
     super();
-    this._nativeShape = PhysicsManager._nativePhysics.createSphereColliderShape(
+    this._nativeShape = PhysicsScene._nativePhysics.createSphereColliderShape(
       this._id,
       this._radius,
       this._material._nativeMaterial
