+++ packages/design/src/physics/index.ts
@@ -1,9 +1,11 @@
-export type { IPhysics } from "./IPhysics";
-export type { IPhysicsManager } from "./IPhysicsManager";
-export type { ICollider } from "./ICollider";
-export type { IStaticCollider } from "./IStaticCollider";
-export type { IDynamicCollider } from "./IDynamicCollider";
-export type { IPhysicsMaterial } from "./IPhysicsMaterial";
 export type { ICharacterController } from "./ICharacterController";
-export * from "./shape/index";
+export type { ICollider } from "./ICollider";
+export type { IDynamicCollider } from "./IDynamicCollider";
+export type { IPhysics } from "./IPhysics";
+export type { IPhysicsMaterial } from "./IPhysicsMaterial";
+export type { IPhysicsScene } from "./IPhysicsScene";
+export type { IPhysicsManager } from "./IPhysicsManager";
+export type { IStaticCollider } from "./IStaticCollider";
 export * from "./joints/index";
+export * from "./shape/index";
+
