+++ tests/src/core/physics/PhysicsManager.test.ts
@@ -10,12 +10,12 @@ describe("physics manager test", () => {
     engine = await WebGLEngine.create({ canvas: document.createElement("canvas"), physics: new LitePhysics() });
   });
 
-  it("constructor", async () => {
+  it("constructor", () => {
     expect(engine.physicsManager.gravity.y).to.eq(-9.81);
     expect(engine.physicsManager.fixedTimeStep).to.eq(1 / 60);
   });
 
-  it("raycast", async () => {
+  it("raycast", () => {
     const scene = engine.sceneManager.activeScene;
     const rootEntity = scene.createRootEntity("root");
